requirements.txt
mnemonic
pycryptodome
cryptography